+++ app/commands/MakerCommand.php
@@ -5,6 +5,16 @@ use System\Database\{MyPDO, MyQuery};
 
 class MakerCommand extends Command
 {
+
+  public static array $command = array(
+    [
+      "cmd"       => "make",
+      "mode"      => "start",
+      "class"     => MakerCommand::class,
+      "fn"        => "switcher",
+    ],
+  );
+
   public function switcher()
   {
     // get category command