+++ app/core/CLI.php
@@ -2,24 +2,40 @@
 
 namespace System\Apps;
 
+use CronCommand;
+use HelpCommand;
 use Helper\String\Str;
+use MakerCommand;
+use ServeCommand;
 
 class CLI
 {
+  public static $command = [];
+
   public function __construct(array $arguments)
   {
     // handle commad empty
     $baseArgs = $arguments[1] ?? '--help';
 
     // load register command
-    $commands = COMMAND_CONFIG;
+    self::$command = array_merge(
+      // help command
+      HelpCommand::$command,
+      // make somthink command
+      MakerCommand::$command,
+      // server command
+      ServeCommand::$command,
+      // cron
+      CronCommand::$command,
+      // more command here
+    );
 
-    foreach ($commands as $cmd) {
+    foreach (self::$command as $cmd) {
       // matching alias
       $use = $cmd['cmd'];
       if (is_array($use)) {
-        foreach ($use as $alias) {
-          $valid = $this->cekAlias($alias, $baseArgs, $cmd['mode']);
+        foreach ($use as $cmd_alias) {
+          $valid = $this->cekAlias($cmd_alias, $baseArgs, $cmd['mode']);
           if ($valid) {
             break;
           }