+++ app/core/Config.php
@@ -15,12 +15,10 @@ class Config
     // load config
     $app_config = include(dirname(__DIR__, 2) . '/app/config/app.config.php');
     $dbs_config = include(dirname(__DIR__, 2) . '/app/config/database.config.php');
-    $command_config = include(dirname(__DIR__, 2) . '/app/config/command.config.php');
 
     // excute config
     $this->appConfig($app_config);
     $this->databeseConfig($dbs_config);
-    $this->commandConfig($command_config);
   }
 
   private function appConfig(array $config)
@@ -52,10 +50,5 @@ class Config
     define('DB_USER', $config['DB_USER']);
     define('DB_PASS', $config['DB_PASS']);
     define('DB_NAME', $config['DB_NAME']);
-  }
-
-  private function commandConfig(array $config)
-  {
-    define('COMMAND_CONFIG', $config);
   }
 }